lxcfs, bindings: set fd_hierarchies in constructor
We need to initialize fd_hierarchies in the __constructor__ to guarantee a
consistent view across users of the shared library. Note well that hierarchies
and fd_hierarchies are considered read-only after the __constructor__ has run.
